refactor(user): move user model into user app
