+++ tests/auth/test_register.py
@@ -4,7 +4,7 @@ from http import HTTPStatus
 from flask import Flask
 from flask.testing import FlaskClient
 
-from aphorism.apps.auth.model import User
+from aphorism.apps.user.model import User
 from tests.conftest import register_user
